[fix] Test if config aleardy enabled

+++ .install/2_Install_app.sh
@@ -9,10 +9,14 @@ pause
 
 # Apache2
 echo "Configuring Apache2 ..."
-sudo cp raph_bot_virtualhost.conf /etc/apache2/sites-available/raph_bot.conf
-sudo nano /etc/apache2/sites-available/raph_bot.conf
-sudo a2ensite raph_bot.conf
-sudo systemctl reload apache2
+if [ -f "/etc/apache2/sites-enabled/raph_bot.conf" ]; then
+    echo -e "\t- Core config file already exist."
+else 
+    sudo cp raph_bot_virtualhost.conf /etc/apache2/sites-available/raph_bot.conf
+    sudo nano /etc/apache2/sites-available/raph_bot.conf
+    sudo a2ensite raph_bot.conf
+    sudo systemctl reload apache2
+fi
 
 # Log File
 echo -e "\nCreating log files"
